requirements.txt
absl-py==2.1.0
altair==4.1.0
astunparse==1.6.3
attrs==24.2.0
backports.zoneinfo==0.2.1
blinker==1.6.3
cachetools==5.5.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
decorator==5.1.1
entrypoints==0.4
flatbuffers==1.12
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.3
h5py==3.8.0
idna==3.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
Jinja2==3.1.4
joblib==1.3.2
jsonschema==4.17.3
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==20.9
pandas==1.3.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.0
pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydeck==0.8.1b1
Pygments==2.17.2
Pympler==1.1
pyparsing==3.1.4
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
pytz==2024.2
pytz-deprecation-shim==0.1.0.post0
pywin32==308
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.8.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
semver==3.0.2
six==1.16.0
smmap==5.0.1
streamlit==1.17.0
tenacity==8.2.3
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.0
tensorflow-estimator==2.9.0
tensorflow-hub==0.12.0
tensorflow-intel==2.11.1
tensorflow-io-gcs-filesystem==0.31.0
tensorflowjs==3.19.0
termcolor==2.3.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.1
tornado==6.2
tqdm==4.66.5
typing_extensions==4.7.1
tzdata==2024.2
tzlocal==4.3.1
urllib3==2.0.7
validators==0.20.0
watchdog==3.0.0
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.16.0
zipp==3.15.0
